Break api_external into files, rearrange
