+++ crates/holochain/src/conductor/api/api_external/app_interface.rs
@@ -1,0 +1,121 @@
+use super::InterfaceApi;
+use crate::conductor::api::error::{ConductorApiResult, ExternalApiWireError, SerializationError};
+use crate::core::ribosome::{ZomeCallInvocation, ZomeCallInvocationResponse};
+use crate::{
+    conductor::{
+        interface::error::{InterfaceError, InterfaceResult},
+        ConductorHandle,
+    },
+    core::workflow::ZomeCallInvocationResult,
+};
+use holochain_serialized_bytes::prelude::*;
+
+/// The interface that a Conductor exposes to the outside world.
+#[async_trait::async_trait]
+pub trait AppInterfaceApi: 'static + Send + Sync + Clone {
+    /// Invoke a zome function on any cell in this conductor.
+    async fn call_zome(
+        &self,
+        invocation: ZomeCallInvocation,
+    ) -> ConductorApiResult<ZomeCallInvocationResult>;
+
+    // -- provided -- //
+
+    /// Routes the [AppRequest] to the [AppResponse]
+    async fn handle_request(&self, request: AppRequest) -> AppResponse {
+        let res: ConductorApiResult<AppResponse> = async move {
+            match request {
+                AppRequest::ZomeCallInvocationRequest(request) => {
+                    match self.call_zome(*request).await? {
+                        Ok(response) => Ok(AppResponse::ZomeCallInvocationResponse {
+                            response: Box::new(response),
+                        }),
+                        Err(e) => Ok(AppResponse::Error(e.into())),
+                    }
+                }
+                _ => unimplemented!(),
+            }
+        }
+        .await;
+
+        match res {
+            Ok(response) => response,
+            Err(e) => AppResponse::Error(e.into()),
+        }
+    }
+}
+
+/// The Conductor lives inside an Arc<RwLock<_>> which is shared with all
+/// other Api references
+#[derive(Clone)]
+pub struct RealAppInterfaceApi {
+    conductor_handle: ConductorHandle,
+}
+
+impl RealAppInterfaceApi {
+    /// Create a new instance from a shared Conductor reference
+    pub fn new(conductor_handle: ConductorHandle) -> Self {
+        Self { conductor_handle }
+    }
+}
+
+#[async_trait::async_trait]
+impl AppInterfaceApi for RealAppInterfaceApi {
+    async fn call_zome(
+        &self,
+        invocation: ZomeCallInvocation,
+    ) -> ConductorApiResult<ZomeCallInvocationResult> {
+        self.conductor_handle.call_zome(invocation).await
+    }
+}
+
+#[async_trait::async_trait]
+impl InterfaceApi for RealAppInterfaceApi {
+    type ApiRequest = AppRequest;
+    type ApiResponse = AppResponse;
+    async fn handle_request(
+        &self,
+        request: Result<Self::ApiRequest, SerializedBytesError>,
+    ) -> InterfaceResult<Self::ApiResponse> {
+        self.conductor_handle
+            .check_running()
+            .await
+            .map_err(InterfaceError::RequestHandler)?;
+        match request {
+            Ok(request) => Ok(AppInterfaceApi::handle_request(self, request).await),
+            Err(e) => Ok(AppResponse::Error(SerializationError::from(e).into())),
+        }
+    }
+}
+
+/// The set of messages that a conductor understands how to handle over an App interface
+#[derive(Clone, Debug, serde::Serialize, serde::Deserialize, SerializedBytes)]
+#[serde(rename = "snake-case", tag = "type", content = "data")]
+pub enum AppRequest {
+    /// Asks the conductor to do some crypto
+    CryptoRequest(Box<CryptoRequest>),
+    /// Call a zome function
+    ZomeCallInvocationRequest(Box<ZomeCallInvocation>),
+}
+
+/// Responses to requests received on an App interface
+#[derive(Clone, Debug, serde::Serialize, serde::Deserialize, SerializedBytes)]
+#[serde(rename = "snake-case", tag = "type", content = "data")]
+pub enum AppResponse {
+    /// There has been an error in the request
+    Error(ExternalApiWireError),
+    /// The response to a zome call
+    ZomeCallInvocationResponse {
+        /// The data that was returned by this call
+        response: Box<ZomeCallInvocationResponse>,
+    },
+}
+
+#[allow(missing_docs)]
+#[derive(Clone, Debug, serde::Serialize, serde::Deserialize)]
+#[serde(rename = "snake-case", tag = "type", content = "data")]
+pub enum CryptoRequest {
+    Sign(String),
+    Decrypt(String),
+    Encrypt(String),
+}
